correcting issues

+++ Services/Services/CustomerService.cs
@@ -16,23 +16,25 @@ public class CustomerService : ICustomerService
         return await _repository.GetAllAsync();
     }
 
-    public async Task<Customer?> GetCustomerByIdAsync(Guid id)
+    public async Task<Customer?> GetCustomerByIdAsync(Guid? id)
     {
-        return await _repository.GetByIdAsync(id);
+        if(id == null) return null;
+        return await _repository.GetByIdAsync((Guid)id);
     }
 
     public async Task<Customer> AddCustomerAsync(Customer customer)
     {
         return await _repository.AddAsync(customer);
     }
-    public async Task<> UpdateCustomerAsync(Customer customer)
+    public async Task UpdateCustomerAsync(Customer customer)
     {
-        return await _repository.UpdateAsync(customer);
+         await _repository.UpdateAsync(customer);
     }
-    public async Task<Customer> DeleteCustomerAsync(Guid id);
+    public async Task DeleteCustomerAsync(Guid id)
     {
         var customer = await _repository.GetByIdAsync(id);
-        return await _repository.DeleteAsync(customer);
+        if(customer ==null) return;
+        await _repository.DeleteAsync(customer);
     }
