+++ Services/Services/GenreService.cs
@@ -1,8 +1,12 @@
+using MvcMovieRepo.Models;
+using MvcMovieRepo.Repositories;
+using mvcMovieRepositoryDotnet.Services.ServiceContracts;
+
 public class GenreService : IGenreService
 {
-    private readonly IGenericeRepository<Genre> _repository;
+    private readonly GenreRepository _repository;
 
-    public GenreService(IGenericeRepository<Genre> repository)
+    public GenreService(GenreRepository repository)
     {
         _repository = repository;
     }
@@ -12,9 +16,10 @@ public class GenreService : IGenreService
         return await _repository.GetAllAsync();
     }
 
-    public async Task<Genre?> GetGenreByIdAsync(Guid id)
+    public async Task<Genre?> GetGenreByIdAsync(Guid? id)
     {
-        return await _repository.GetByIdAsync(id);
+        if(id==null) return null;
+        return await _repository.GetByIdAsync((Guid)id);
     }
 
     public async Task<Genre> AddGenreAsync(Genre genre)
@@ -22,14 +27,15 @@ public class GenreService : IGenreService
         return await _repository.AddAsync(genre);
 
     }
-    public async Task<Genre> UpdateGenreAsync(Genre genre)
+    public async Task UpdateGenreAsync(Genre genre)
     {
-        return await _repository.UpdateAsync(genre);
+        await _repository.UpdateAsync(genre);
 
     }
-    public async Task<Genre> DeleteGenreAsync(Guid id)
+    public async Task DeleteGenreAsync(Guid id){
         var genre = await _repository.GetByIdAsync(id);
-        return await _repository.DeleteAsync(genre);
+        if(genre != null)
+            await _repository.DeleteAsync(genre);
     }
